requirements.txt
gradio
torch
torchvision
transformers
opencv-python-headless
numpy
Pillow
huggingface-hub
pyngrok
onnxruntime-gpu